Initial commit
Will launch and make a REPL available.

The :dbmunge profile will be folded into project.clj.

I unfortunately cannot fold in :datomicdb since it has my
username/password for Cognitect's/Datomic.com's repository server and I
would be in violation of my license terms were I to redistribute the
Datomic Pro .jar elsewhere.  Sign up for a Datomic Pro Starter license
at https://my.datomic.com and check it out!  Datomic is weird and
proprietary and looks like it comes out of the 1970s (hierarchically
structured schema-oriented non-relational database transactor that
prefers to use someone else's backend?  What?) but also awesome and fast
and immutable and likes to see things as changes to a core dataset
rather than a blob that one hopes looks like where things are at *right
now*, the past or data not explicitly asked for be damned.

Realistically, I would advocate for using a more tested open source DB
(or MSSQL/Oracle if you insist on a closed-source DB server), and
optimizing at the client end or with various schema hacks when your
backend falls short of keeping your data safe, sane and catalogued in a
way that a human brain can parse.  In a sense, this is what Datomic
does, but it makes decisions for you and has a design (albeit a modern,
advanced and well-thought-out one) that most RDBMS developers abandoned
around 1980.  Also, $3000-$16000 a *year* (no, it's not a one-time fee,
you're subscribing to it a la Adobe CC, MSDN, Office365, what-have-you)
is a lot for a DBMS that primarily supports Java, has a Clojure wrapper,
and requires that any other JVM-hosted language use the Java API
(thankfully nearly 100% of them support Java interop, or support of
Datomic would be impossible).  Use Python, Ruby, PHP or Perl and want to
have a datomic peer running as part of your application?  There's a peer
REST API, and guess what, you need to have a JRE installed on the
deployment platform and permission from your application hosting
provider to have a Java process running 24/7.

So yeah, Datomic is neat, it's written in part by the guy who wrote the
majority of Clojure 1.0, and manages to be one of the leading FP
trendsetters while having an academic grounding in C++, plus a bunch of
other lead Clojure devs, but given that it's written entirely in Java,
has an oddly OO mindset, is closed-source, looks like what you would get
if you asked the leading DB developers and DBAs from 1969 to design a
cutting-edge database, and costs $3000 a seat minimum, I find it hard to
recommend to people interested in whatever's this year's Hip Database
Tech, open source platforms, functional programming, or pure OO.  It's
just too damned weird.